move `wrap_key_data` to `jax.random`
This is a fine function for the public API, rather than `jax.extend`.
