--- jax/_src/extend/random.py
@@ -1,23 +1,0 @@
-# Copyright 2023 The JAX Authors.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     https://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-from typing import Optional
-
-from jax._src import prng
-from jax._src import random
-from jax._src.typing import Array
-
-def wrap_key_data(key_bits_array: Array, *, impl: Optional[str] = None):
-  impl_obj = random.resolve_prng_impl(impl)
-  return prng.random_wrap(key_bits_array, impl=impl_obj)
